refactor: CodefClient 인터페이스 추출 및 구현체를 CodefRealClient로 분리
- 기존 CodefClient 클래스를 인터페이스로 바꾸고, 실제 HTTP 호출 로직은 CodefRealClient로 옮겼습니다. 이렇게 추상화를 해줘야 테스트 프로필에서 Mock 구현체로 쉽게 교체 가능합니다.
